Added php to pull from database
HAVE NOT PROPERLY TESTED THIS. Creates a php session to save the quiz questions and answers, populates quiz area with one and removes it from the array. Idea is that the check will run, display whether answer is correct, then reload the page with the next question until the session array is empty, then retrying the quiz will just end the session and start a new one.