.js for the video display
Made specifically for the beginner page, but can work with others. Made a single function to check and display all videos, keeps track of each index separately despite the same function being called. Look at the begin.php file for calls to this .js to understand what needs to be passed